correct README examples
