+++ .gitignore
@@ -26,3 +26,4 @@ target/
 
 # @tuana9a
 *tmp*
+.vscode/
